fix: double render in studio mode
