+++ src/renderer.ts
@@ -25,22 +25,6 @@
  *  });
  * ```
  */
-import { createRoot } from 'react-dom/client';
-import App from './components/app';
-import './i18n';
-import './index.css';
-import { DocumentEditorDebugTools } from './utils/document-editor-debug-tools';
-import serviceWorkerManager from './utils/registerServiceWorker';
 
-const AppElement = document.getElementById('app');
-if (!AppElement) {
-    throw new Error('#app element not found!');
-}
-
-if (import.meta.env.MODE === 'development') {
-    window.documentEditorDebugTools = new DocumentEditorDebugTools();
-} else {
-    serviceWorkerManager.registerServiceWorker(false);
-}
-
-createRoot(AppElement).render(App);
+// splits app code into separate chunk to avoid circular dependencies
+import('./index');
